Add files via upload
Plots of Histograms for Degree Distributions named HistTagDegrees are histograms of Degree Distributions for Tags where the upper bound is 10000 in case of 10k2, and the lower bound is 3, meaning it is a histogram of Tags with degrees BETWEEN 3 and 10000. Same with Stories; however, we shouldn't worry about histograms of Degree Distributions of Stories too much. The main purpose of Degree Distributions for Tags is to highlight that it follows a power law, even when the Maximum Degree was constrained to 10, and the minimum was 3. Therefore, these figures act as supporting evidence to remove Tags with extremely high or low Tags. 
!!! IMPORTANT: I tried removing Tags by the quantile function, and setting it to 99% Percentile; however, it removes too many useful tags. 
The Hidden Gem plots are a result of either Katz or Eigen centrality measures to recommend non popular Stories that have high centrality scores.